auditbeat: Split common tty definitions
No functional changes, just move tty things to a common place so that the
upcoming system/process provider can use it.

Last of the PRs to split common code for the new system/process provider (I
promise).
